+++ .cargo/config.toml
@@ -1,0 +1,3 @@
+[alias]
+xtest = "run --package test-runner --bin test-runner --"
+xt = "xtest"
