+++ tests/runner/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "trun"
+name = "test-runner"
 version = "0.1.0"
 edition = "2018"
